feat!: Created a .NET API project as a backend and ReactJS as a frontend.
The backend part is more or less finalized. Frontend is left in boilerplate stage.

feat!: Implemented ForcastWeather class as a serialization model as well as sub-classes for different weather components

feat!: Created an API weather controller for interaction with the frontend.

docs: Worte page for introduction and brief API instruction.